Moved unpack_fieldsfile to conftest.py. Suggested by @atteggiani
