+++ test/drivers/test_esm1p6_filenames.py
@@ -6,34 +6,11 @@ import numpy as np
 import os
 from pathlib import Path
 import re
-import tarfile
 import iris
 from iris.cube import Cube
 
 from um2nc.drivers.esm1p6 import Esm1p6Driver, Esm1p6DelayedCubePath
 from test_esm1p6 import mock_atmosphere_dir
-
-
-# FIXME: This should be imported from test_single_field not duplicated
-#  or some common package that both can see
-@pytest.fixture
-def unpack_fieldsfile(tmp_path):
-    """
-    A fieldfile with all-zero fields has been tar'ed up and compressed with lzma
-    in the expected directory structure.
-    zeroed_fieldsfile.tar.lama
-    └── atmosphere
-        ├── aiihca.pa01apr # This fieldsfile has all the fields set to zero
-        └── xhist # um2nc needs this file too
-
-    Compressed this file is 116kB, uncompressed it is 648MB.
-    """
-    src = "test/data/zeroed_fieldsfile.tar.lama"
-    dst = tmp_path / "zeroed_fieldsfile"
-    tar = tarfile.open(src)
-    tar.extractall(dst, filter="data")
-
-    return dst
 
 
 @pytest.mark.parametrize(
